Fix issue where pool serializer would not be same as worker
The issue was encountered when initializing redis through RedisSettings only
instead of explicitly setting the arq connection pool.

If cron was used and `job_serializer` was not `pickle` the cron item would be
serialized with pickle.dumps and then attempted to deserialize with the configured
`job_deserializer` in worker, causing an error.

Quick workaround is to create the pool yourself, otherwise always use the
pickle serializer.