Allow user to configure a default job expiry length
Previously, expires_extra_ms in arq/connections.py set a default expiry
of 1 day from when a job was expected to start to when it would expire
if no expiry was set in enqueue_job().

This commit keeps that as the default, but allows both ArqRedis and Worker
to have a larger default expiry configured, as self.expires_extra_ms.

expires_extra_ms is also moved to arq/constants.py along with other
shared constants.