Add tests for --custom-log-dict CLI option
Includes a workaround for deprecation warnings for asyncio.get_event_loop()
(as described in https://github.com/pytest-dev/pytest-asyncio/issues/212),
following the mocking out done in https://github.com/samuelcolvin/arq/pull/298.

Also, as requested in review, use import_string to handle the import, and
if it's an invalid import path than let it raise an ImportError to inform
the user.