Fix race condition causing status not_found
Checking job status consited of three separate calls to Redis,
where the first one checked if result is present and the second
checked if job is in progress. If a job got completed between
these two calls Job.status() would erroneously return status
not_found.

The solution is to make all the calls to Redis in one transaction.