Raise ResultNotFound when Job.result() finds no job and no result
Job.result() assumed that a result will eventually show up in Redis.
This assumption does not hold for jobs which are not in the queue
and jobs which do not keep result. In such cases Job.result() would
hang forever (or until the optional timeout is reached).

This PR changes Job.result() so that it raises a new exception
ResultNotFound, rather than waiting forever, if result is not found
and the job is not in the queue.