Add opentelemetry integration
Add a draft opentelemtry integration consisting of a span propagator, a
decorator for instrumenting jobs and a class for hooking into
enqueue_job with span propagation. Additionally there are some metrics
for the number of enqueued jobs.

This is a reasonably bare-bones implementation that needs work in a few
areas which I will can come back to, but I think there are a few
weaknesses I can think of:

* Having to explicitly instrument jobs using the decorator is a bit
  clunky and perhaps it should happen automatically?
* `InstrumentedArqRedis` is a bit awkward to use and differs quite a lot
  from the `opentelemetry-contrib` modules which use Instrumentors.
* I'm not certain about whether this should be an extra to the main
  `arq` package or a separate extension.