Uses `testcontainers` to provide a redis instance for the unit tests
We can't assume that a dev would want the unit tests to connect to the locally
running redis at `localhost:6379`, so I'd rather shift that to assuming that the
user has a Docker-compatible system available.  This makes the test suite easier
to reason about and lets us control the redis version with more precision, and
it makes the initial startup experience easier for new devs.